+++ HISTORY.rst
@@ -2,6 +2,25 @@
 History
 =======
 
+0.2.0 (August 29th, 2018)
+=========================
+
+Big changes:
+
+* Siggen is re-united with Socorro's signature generation system. Generally
+  we'll make changes in the Socorro repository and then copy them here.
+
+Bug fixes:
+
+* 1477726: add ``std:alloc::rust_oom`` to prefix list
+* 1481282: rework frame normalization so it treats C/C++ frames differently
+  than Rust frames
+* 1477013: rewrite collapse to correctly handle Rust trait methods
+* 1478383: drop prefix and return type in function signatures; add handling
+  for "const" in function signatures
+* 1306643: document signature generation pipeline
+
+
 0.1.3 (August 3rd, 2018)
 ========================